+++ ITI.CEI40.Monitor/Models/View Models/EngineerChrtViewModel.cs
@@ -8,8 +8,8 @@ namespace ITI.CEI40.Monitor.Models.View_Models
     public class EngineerChrtViewModel
     {
         public List<string> Months { get; set; }
-        public List<int> Quality { get; set; }
-        public List<int> Time { get; set; }
+        public List<float> Quality { get; set; }
+        public List<float> Time { get; set; }
         public string EngineerName { get; set; }
         public List<int> SubMonth { get; set; }
     }
